% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_file_types.R
\name{get_all_file_types}
\alias{get_all_file_types}
\title{Get all file types in the Kirby21.mricloud Package}
\usage{
get_all_file_types()
}
\value{
Vector of character types
}
\description{
Return the file types for the people scanned available in the 
kirby21.mricloud package
}
