% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\title{Get IDs with Data in the Kirby21.mricloud Package}
\usage{
get_ids()
}
\value{
Vector of numeric ids
}
\description{
Return the IDs for the people scanned available in the 
kirby21.mricloud package
}
